requirements.txt
git+https://github.com/facebookresearch/ImageBind.git
git+https://github.com/openai/CLIP.git
ipython
pandas
Pillow
requests
seaborn
torch
openai
anthropic
transformers